[Mission01][Feedback] 파일 분리 적용 / live-server 설치
