+++ mission001/YongHoonJJo/index.html
@@ -1,61 +1,41 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>이벤트 - TODOS</title>
-  <link rel="stylesheet" href="css/style.css">
-</head>
-<body>
-<section class="todoapp">
-  <div>
-    <h1>TODOS</h1>
-    <input id="new-todo-title" class="new-todo" placeholder="할일을 추가해주세요" autofocus>
-  </div>
-  <div class="main">
-    <input class="toggle-all" type="checkbox">
-    <ul id="todo-list" class="todo-list">
-      <li>
-        <div class="view">
-          <input class="toggle" type="checkbox">
-          <label class="label">새로운 타이틀</label>
-          <button class="destroy"></button>
-        </div>
-        <input class="edit" value="새로운 타이틀">
-      </li>
-      <li class="editing">
-        <div class="view">
-          <input class="toggle" type="checkbox">
-          <label class="label">완료된 타이틀</label>
-          <button class="destroy"></button>
-        </div>
-        <input class="edit" value="완료된 타이틀">
-      </li>
-      <li class="completed">
-        <div class="view">
-          <input class="toggle" type="checkbox">
-          <label class="label">완료된 타이틀</label>
-          <button class="destroy"></button>
-        </div>
-        <input class="edit" value="완료된 타이틀">
-      </li>
-    </ul>
-  </div>
-  <div class="count-container">
-    <span class="todo-count">총 <strong>0</strong> 개</span>
-    <ul class="filters">
-      <li>
-        <a class="all selected" href="#/">전체보기</a>
-      </li>
-      <li>
-        <a class="active" href="#/active">해야할 일</a>
-      </li>
-      <li>
-        <a class="completed" href="#/completed">완료한 일</a>
-      </li>
-    </ul>
-  </div>
-</section>
-<script src="js/app.js"></script>
-</body>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>이벤트 - TODOS</title>
+    <link rel="stylesheet" href="css/style.css" />
+  </head>
+  <body>
+    <section class="todoapp">
+      <div>
+        <h1>TODOS</h1>
+        <input
+          id="new-todo-title"
+          class="new-todo"
+          placeholder="할일을 추가해주세요"
+          autofocus
+        />
+      </div>
+      <div class="main">
+        <input class="toggle-all" type="checkbox" />
+        <ul id="todo-list" class="todo-list"></ul>
+      </div>
+      <div class="count-container">
+        <span class="todo-count">총 <strong>0</strong> 개</span>
+        <ul class="filters">
+          <li>
+            <a class="all selected" href="#/">전체보기</a>
+          </li>
+          <li>
+            <a class="active" href="#/active">해야할 일</a>
+          </li>
+          <li>
+            <a class="completed" href="#/completed">완료한 일</a>
+          </li>
+        </ul>
+      </div>
+    </section>
+    <script type="module" src="js/App.js"></script>
+  </body>
 </html>
